To cut a long storty short I've upgraded to Boost 1.33. This breaks the uBLAS co de I originally had so these commits basically fix compiler errors caused by uBL AS changing from 1.32 to 1.33. The is starting to piss me off ... they keep chan ging stuff that breaks my code.........

+++ trunk/Physics/FixedPlane.hpp
@@ -22,7 +22,7 @@ namespace Edge
 		virtual void CalculateDJacobian();
 		virtual int GetRowSize() const;
 		void SetMatrixCol(int Col) {m_MatrixCol = Col;}
-		void SetNormal(const bnu::vector<int>& Normal);
+		void SetNormal(const bnu::vector<double>& Normal);
 		const bnu::vector<double>& GetNormal();
 		virtual void ResizeJacobian(int NumParticles);
 		virtual void ResizeDJacobian(int NumParticles);
